Created the ability for you to consume or drop items from your inventory.
I also realized I was never saving anything from the Character Scene like I was from the Weapons R Us Scene, so I fixed that.

I lowered the amount of gold and value of items able to be found in the Fields.

I changed the TextEdit on the Battle Scene to be a RichTextLabel so it will scroll to the end more easily.