Dying now has more of a penalty.
When you die from an animal, your equipment takes damage. When you die to a human or boss, you lose 5-10% of your gold on hand, and up to one item of 200 or lesser value.

I also removed the default weapon/armor on the Hero's Unequip method since that is no longer being used for the same purpose as it was in my WPF version of the game.