Implemented Chapel, change health to bars
I also fixed a text error when attempting to equip an item of a too high level.

I changed the way I call my DisplayPopup() method, by putting it in GameState and resetting the size to the default on each display of it so that it isn't gigantic after a long message in it previously.